ResultSheet: move modal sheet composable here
