+++ app/src/main/java/page/ooooo/geoshare/ui/components/ResultSheet.kt
@@ -1,13 +1,22 @@
 package page.ooooo.geoshare.ui.components
 
+import android.annotation.SuppressLint
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.safeDrawing
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.Surface
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SheetValue
+import androidx.compose.material3.rememberBottomSheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
@@ -17,6 +26,7 @@ import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.persistentListOf
+import kotlinx.coroutines.launch
 import page.ooooo.geoshare.R
 import page.ooooo.geoshare.data.OutputRepository
 import page.ooooo.geoshare.data.di.defaultFakeLinks
@@ -32,57 +42,79 @@ import page.ooooo.geoshare.lib.outputs.PointsOutput
 import page.ooooo.geoshare.ui.theme.AppTheme
 import page.ooooo.geoshare.ui.theme.LocalSpacing
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ResultSheet(
     points: Points,
     selectedPointIndex: Int,
     appDetails: AppDetails,
+    initialValue: SheetValue = SheetValue.Hidden,
     outputsForPoint: List<PointOutput>,
     outputsForPoints: List<PointsOutput>,
     onExecute: (action: Action<*>) -> Unit,
-    onHide: () -> Unit,
+    onSelectPointIndex: (index: Int?) -> Unit,
 ) {
+    val coroutineScope = rememberCoroutineScope()
+    val sheetState = rememberBottomSheetState(initialValue)
+
     val selectedPoint = points.getOrNull(selectedPointIndex) ?: return
 
-    LazyColumn(
-        Modifier
-            .semantics { testTagsAsResourceId = true }
-            .testTag("geoShareResultSheet"),
+    fun hide() {
+        coroutineScope.launch { sheetState.hide() }.invokeOnCompletion {
+            if (!sheetState.isVisible) {
+                onSelectPointIndex(null)
+            }
+        }
+    }
+
+    ModalBottomSheet(
+        onDismissRequest = { onSelectPointIndex(null) },
+        modifier = Modifier
+            // Set and consume insets to prevent unclickable items when the sheet is expanded (probably a bug in
+            // Compose Material 3)
+            .windowInsetsPadding(WindowInsets.safeDrawing),
+        sheetState = sheetState,
     ) {
-        item {
-            ResultSuccessSheetItemGroup(
-                title = if (points.size > 1) {
-                    stringResource(R.string.conversion_succeeded_point_number, selectedPointIndex + 1)
-                } else {
-                    null
-                },
-                appDetails = appDetails,
-                actions = outputsForPoint.map { it.toAction(selectedPoint) },
-                value = selectedPoint,
-                onClick = { action ->
-                    onHide()
-                    onExecute(action)
-                },
-            )
-        }
-        item {
-            Spacer(Modifier.height(LocalSpacing.current.mediumAdaptive))
-        }
-        item {
-            ResultSuccessSheetItemGroup(
-                title = if (points.size > 1) {
-                    stringResource(R.string.conversion_succeeded_point_all, points.size)
-                } else {
-                    null
-                },
-                appDetails = appDetails,
-                actions = outputsForPoints.map { it.toAction(points) },
-                value = points,
-                onClick = { action ->
-                    onHide()
-                    onExecute(action)
-                },
-            )
+        LazyColumn(
+            Modifier
+                .semantics { testTagsAsResourceId = true }
+                .testTag("geoShareResultSheet"),
+        ) {
+            item {
+                ResultSuccessSheetItemGroup(
+                    title = if (points.size > 1) {
+                        stringResource(R.string.conversion_succeeded_point_number, selectedPointIndex + 1)
+                    } else {
+                        null
+                    },
+                    appDetails = appDetails,
+                    actions = outputsForPoint.map { it.toAction(selectedPoint) },
+                    value = selectedPoint,
+                    onClick = { action ->
+                        hide()
+                        onExecute(action)
+                    },
+                )
+            }
+            item {
+                Spacer(Modifier.height(LocalSpacing.current.mediumAdaptive))
+            }
+            item {
+                ResultSuccessSheetItemGroup(
+                    title = if (points.size > 1) {
+                        stringResource(R.string.conversion_succeeded_point_all, points.size)
+                    } else {
+                        null
+                    },
+                    appDetails = appDetails,
+                    actions = outputsForPoints.map { it.toAction(points) },
+                    value = points,
+                    onClick = { action ->
+                        hide()
+                        onExecute(action)
+                    },
+                )
+            }
         }
     }
 }
@@ -116,11 +148,13 @@ private fun <T> ResultSuccessSheetItemGroup(
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Preview(showBackground = true, device = "spec:width=1080px,height=3200px,dpi=440")
 @Composable
 private fun DefaultPreview() {
     AppTheme {
-        Surface {
+        @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+        Scaffold {
             val context = LocalContext.current
             val geometries = Geometries(context)
             val coordinateConverter = CoordinateConverter(geometries)
@@ -131,15 +165,17 @@ private fun DefaultPreview() {
                 points = persistentListOf(WGS84Point(NaivePoint.example), WGS84Point(NaivePoint.genRandomPoint())),
                 selectedPointIndex = 1,
                 appDetails = emptyMap(),
+                initialValue = SheetValue.Expanded,
                 outputsForPoint = outputRepository.getOutputsForPoint(defaultFakeLinks),
                 outputsForPoints = outputRepository.getOutputsForPoints(),
-                onHide = {},
                 onExecute = {},
+                onSelectPointIndex = {},
             )
         }
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Preview(
     showBackground = true,
     device = "spec:width=1080px,height=3200px,dpi=440",
@@ -148,7 +184,8 @@ private fun DefaultPreview() {
 @Composable
 private fun DarkPreview() {
     AppTheme {
-        Surface {
+        @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+        Scaffold {
             val context = LocalContext.current
             val geometries = Geometries(context)
             val coordinateConverter = CoordinateConverter(geometries)
@@ -159,20 +196,23 @@ private fun DarkPreview() {
                 points = persistentListOf(WGS84Point(NaivePoint.example), WGS84Point(NaivePoint.genRandomPoint())),
                 selectedPointIndex = 1,
                 appDetails = emptyMap(),
+                initialValue = SheetValue.Expanded,
                 outputsForPoint = outputRepository.getOutputsForPoint(defaultFakeLinks),
                 outputsForPoints = outputRepository.getOutputsForPoints(),
-                onHide = {},
                 onExecute = {},
+                onSelectPointIndex = {},
             )
         }
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Preview(showBackground = true, device = "spec:width=1080px,height=3200px,dpi=440")
 @Composable
 private fun LastPointPreview() {
     AppTheme {
-        Surface {
+        @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+        Scaffold {
             val context = LocalContext.current
             val geometries = Geometries(context)
             val coordinateConverter = CoordinateConverter(geometries)
@@ -183,15 +223,17 @@ private fun LastPointPreview() {
                 points = persistentListOf(WGS84Point(NaivePoint.example)),
                 selectedPointIndex = 0,
                 appDetails = emptyMap(),
+                initialValue = SheetValue.Expanded,
                 outputsForPoint = outputRepository.getOutputsForPoint(defaultFakeLinks),
                 outputsForPoints = outputRepository.getOutputsForPoints(),
-                onHide = {},
                 onExecute = {},
+                onSelectPointIndex = {},
             )
         }
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Preview(
     showBackground = true,
     device = "spec:width=1080px,height=3200px,dpi=440",
@@ -200,7 +242,8 @@ private fun LastPointPreview() {
 @Composable
 private fun DarkLastPointPreview() {
     AppTheme {
-        Surface {
+        @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+        Scaffold {
             val context = LocalContext.current
             val geometries = Geometries(context)
             val coordinateConverter = CoordinateConverter(geometries)
@@ -211,10 +254,11 @@ private fun DarkLastPointPreview() {
                 points = persistentListOf(WGS84Point(NaivePoint.example)),
                 selectedPointIndex = 0,
                 appDetails = emptyMap(),
+                initialValue = SheetValue.Expanded,
                 outputsForPoint = outputRepository.getOutputsForPoint(defaultFakeLinks),
                 outputsForPoints = outputRepository.getOutputsForPoints(),
-                onHide = {},
                 onExecute = {},
+                onSelectPointIndex = {},
             )
         }
     }
